backend/requirements.txt
flask
flask-cors
pandas
openpyxl
matplotlib
seaborn
python-docx
